requirements.txt
aiogram==2.25.1
python-dotenv==1.0.1
openai>=1.0.0
google-genai>=0.1.0
notion-client>=1.0.0
